+++ src/api/getOrderListData.ts
@@ -1,13 +1,8 @@
-import { OrderListQueryParams } from "../hooks/useQueryString";
 import type { OrderData } from "../types/OrderData";
+import { OrderListQueryParams } from "../types/QueryParams";
 import { formatDate } from "../utils/formatDate";
 
-export interface GetOrderListDataProps extends OrderListQueryParams {
-  date?: string;
-  itemAmountPerPage?: number;
-}
-
-interface GetOrderListDataRes {
+export interface GetOrderListDataRes {
   data: OrderData[];
   minPage: number;
   maxPage: number;
@@ -23,7 +18,7 @@ export const getOrderListData = async ({
   page = 1,
   sortBy,
   sortOrder,
-}: GetOrderListDataProps): Promise<GetOrderListDataRes> => {
+}: OrderListQueryParams): Promise<GetOrderListDataRes> => {
   const res = await fetch("/mock_data.json");
   if (!res.ok) throw new Error("데이터를 불러오지 못했습니다");
   let data: OrderData[] = await res.json();
